+++ src/fetch/package-info.ts
@@ -1,46 +1,40 @@
-import { pick } from "lodash-es";
-
 import type { PackageManager } from "../types.ts";
 import { execute } from "./execute.ts";
 
 export const getPackageInfo = async (
   packageManager: PackageManager,
   packageName: string,
-): Promise<{ deprecated?: string; releaseTime?: Record<string, string> }> => {
+): Promise<{
+  deprecated?: string;
+  versions?: Record<string, { deprecated?: string; time: string }>;
+}> => {
   const cmd = {
-    berry: `yarn npm info ${packageName} --fields deprecated,time,versions --json`,
-    npm: `npm view ${packageName} deprecated time versions --json`,
-    pnpm: `npm view ${packageName} deprecated time versions --json`,
-    yarn: `yarn info ${packageName} --json`,
-  }[packageManager];
+    berry: "yarn config get npmRegistryServer",
+    npm: "npm config get registry",
+    pnpm: "pnpm config get registry",
+    yarn: "yarn config get registry",
+  }[packageManager].replace(/\/$/, "");
 
-  return execute(cmd).then((stdout) => {
-    if (!stdout) {
-      return {};
-    }
+  const registry = await execute(cmd);
 
-    const json = JSON.parse(stdout) as unknown;
-    switch (packageManager) {
-      case "yarn": {
-        const {
-          data: { deprecated, time, versions },
-        } = json as {
-          data: {
-            deprecated?: string;
-            time: Record<string, string>;
-            versions: string[];
-          };
-        };
-        return { deprecated, releaseTime: pick(time, versions) };
-      }
-      default: {
-        const { deprecated, time, versions } = json as {
-          deprecated?: string;
-          time: Record<string, string>;
-          versions: string[];
-        };
-        return { deprecated, releaseTime: pick(time, versions) };
-      }
-    }
-  });
+  const { deprecated, time, versions } = (await (
+    await fetch(`${registry}/${packageName}`)
+  ).json()) as {
+    deprecated?: string;
+    time: Record<string, string>;
+    versions: Record<string, { deprecated?: string }>;
+  };
+
+  return {
+    deprecated,
+    versions: Object.fromEntries(
+      Object.entries(versions).map(([version, { deprecated }]) => [
+        version,
+        {
+          deprecated,
+          time: time[version] as string,
+        },
+      ]),
+    ),
+  };
 };